project0/requirements.txt
# Core dependencies
openai>=1.3.7
python-docx>=0.8.11
pypdf>=3.17.1
flask>=2.3.2
beautifulsoup4>=4.12.2
spacy>=3.7.2
scikit-learn>=1.3.2
numpy>=1.24.3
requests>=2.31.0
python-dotenv>=1.0.0
flask-wtf>=1.2.1
bootstrap-flask>=2.3.0
uvicorn
# Testing
pytest>=7.4.3

# Spacy language model
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Type checking
typing-extensions>=4.8.0
mypy>=1.7.0